Move backup_token helper to mobile user view
